requirements.txt
opencv-python-headless==4.10.0.84
ninja

fastapi==0.115.6
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
gunicorn==23.0.0
torch
torchvision
python-multipart
